tech(api): moved CombinedCourseBlueprintItem model file in correct folder
