+++ api/src/quest/domain/models/combined-course-blueprints/value-objects/CombinedCourseBlueprintItem.js
